refactor(economizer): consolidate the context economizer into src/modules/economizer/
Pure refactor (no behavior change): the context-economizer subsystem — the reduce
orchestrator, history folding, and tool-output condensation — moves into a single
module directory behind one public facade header.

- Move the 9 subsystem files (context_reduce, context_fold, tool_condense,
  coord_closet, fold_budget, fold_register, fold_recall, episode_seal, task_rail)
  and their headers into src/modules/economizer/ (git renames; content unchanged).
- New facade src/modules/economizer/economizer.h aggregates the 3 public headers;
  external consumers now include ONLY economizer.h (the sub-headers are module-
  internal). ~10 call sites rewired.
- Build wiring: -Imodules/economizer added to C_FLAGS (and the CMake targets'
  include dirs) so the moved headers resolve; CORE_SRCS + the test Rules.mk .o
  paths updated to modules/economizer/; test relative "../headers/<econ>.h"
  includes fixed to resolve via the new include path.

First of the per-module subdirectory moves (user directive: all modules into their
own subdirectory). The stage/orch modules + seam infra (src/pipeline/) follow in a
second PR. Verified: the economizer + fold + ingress + config unit tests build and
pass; all rewired server objects compile.
