+++ src/modules/economizer/context_reduce.c
