+++ src/posix/agent_runtime.c
@@ -30,8 +30,7 @@
 #include "liveness.h"
 #include "provider_cli_adapter.h"
 #include "config.h"
-#include "context_fold.h"
-#include "context_reduce.h"
+#include "economizer.h"
 #include "fold_recall.h"
 #include "dstr.h"
 #include "cJSON.h"
