+++ src/tests/test_task_rail.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "../headers/task_rail.h"
+#include "task_rail.h"
 
 #define PASS(name) printf("  PASS: %s\n", name)
